+++ app/src/main/java/com/example/app21try6/transaction/transactionedit/TransactionEditViewModelFactory.kt
@@ -18,7 +18,7 @@ class TransactionEditViewModelFactory(
     private val transRepo: TransactionsRepository,
     private val discountRepo: DiscountRepository,
     private val application: Application,
-    private val id:Int
+    private val id: Long
 ):ViewModelProvider.Factory{
     @Suppress("unchecked_cast")
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
